+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2023-xx-xx
+
+### Added
+
+- Freelist implementation
+
+### Fixed
+
+- Memory leak caused by extra refcount in `Cons_from_fast`
+
 ## [0.2.0] - 2023-04-17
 
 ### Added